requirements.txt
streamlit
langchain
lancedb
mistralai
sentence-transformers
python-dotenv
groq
langchain-huggingface
